requirements.txt
google-genai>=1.0.0
requests>=2.31.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
rich>=13.7.0
python-dotenv>=1.0.0
